backend/requirements.txt
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
Pillow==9.2.0
psycopg2-binary==2.8.6
python-dotenv==0.21.1